backend/requirements.txt
requests
pandas
sqlite-utils
fastapi
uvicorn[standard]
databases
aiosqlite
yfinance
python-dotenv